Move typology form into geometry form
